+++ CHANGELOG.md
@@ -20,6 +20,7 @@ i.e. `<Major version>.<Minor version>.<Patch version>`
 - Add capability of overwrite SIP response #150
 - Integrate 3rd-party SIP protection service (ClearIP, Youmail..) #147
 - Ability to customize preset variables
+- Handle nft outbound interconnection
 
 ## [v0.7.1] - 2023-12-25
 - Fix UI access layer #121